Add files via upload
DMPplasticityPC creates a .jld2 file with all the data. the way to access the data is described in the "plot" files. 
This code generates the same behavior than the nromal DMP, but adding some plasticity on the place cells centers and width of activity profile. 
However, although the centers of place cells seem to successfully move toward the platform location through time, the widths appear wider around the platform location, possibly due to the fact that around this location the value function estimation is quite steep, therefore any "bad move" from the agent would create a reduction in value, therefore increasing the width of activity.